refactor(#328): move check_row_count_spike to sync_orchestrator.row_count_spikes
SpikeResult dataclass and check_row_count_spike function relocated from
app/services/ops_monitor.py into new app/services/sync_orchestrator/
row_count_spikes.py. ops_monitor.py retains lazy-shim wrappers for both
(not module-level re-exports, to avoid the existing circular import via
sync_orchestrator.__init__ → adapters → ops_monitor). scheduler.py
updated to import directly from the new path. tests/test_ops_monitor.py
continues to work through the shim. New test file covers the new import
path with regression cases.
